feat(HACBS-563): create apply-mapping bundle
This commit removes the component-mapping bundle and creates the apply-mapping
bundle. The apply-mapping bundle is meant to take an application snapshot and
config file that contains a mapping. The task will return the snapshot with the
mapping applied (or the original snapshot if no mapping exists). Unlike the
component-mapping bundle this replaces, this bundle contains just one task
and no pipeline definition.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>