chore(hacbs-1075): update YAML definitions to trigger CI bundle push
This adds a trivial annotation to YAML definitions to force CI.
Normally CI will update bundles only when they have been updated.
For lack of a way to manually trigger the CI, hence the commit.

The trivial update here is adding a "release" tag to the ones
managed by the release team.

Signed-off-by: Jon Disnard <jdisnard@redhat.com>