feat(HACBS-1552): update upload sbom task
Adapt the task to support multiple images on the input.

Also, add a second step to run the upload_sbom script
for each sbom file.

Also, improve pyxisSecret description to mention
the requirement for key and cert.
Signed-off-by: Martin Malina <mmalina@redhat.com>

Also, update image used by the task to the one
with latest upload_sbom script updates.

Signed-off-by: Martin Malina <mmalina@redhat.com>