feat(HACBS-1767): make create-pyxis-image task support multiple images
This commit changes the create-pyxis-image task to support snapshots
with multiple containerImages contained in it. To do this, the skopeo
inspect is done in the create-pyxis-image task instead of in a
standalone task. Because the skopeo-inspect and
prepare-skopeo-image-string tasks were created just to feed into this
task, they are removed as part of this commit.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>