feat(HACBS-2011): task push-snapshot can push a tag with git sha
This commit adds a new parameter addGitShaTag to the push-snapshot task.
If the parameter is true, it will search for a git commit for each
component in the mappedSnapshot and use the commit value as a tag when
pushing the image. As part of this, the push-snapshot task no longer
uses a bash array for json interpretation and instead uses jq.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>