feat(HACBS-1554): add push-sbom-to-pyxis task to push-to-external-registry
This will ensure that sbom components for the image are pushed to Pyxis
as part of this pipeline.

Also, modify push-sbom-to-pyxis task to use mappedSnapshot instead
of image urls as a parameter - we don't have the urls ready
in the pipeline, so we'll have to get them from the snapshot.

Signed-off-by: Martin Malina <mmalina@redhat.com>