feat(HACBS-2184): enable tekton task testing for apply-mapping
This adds a new PR workflow that installs a kind cluster, tekton,
tkn and then runs tests (defined as pipelines) for tasks.

Documentation is added to CONTRIBUTING.md.

Two test pipelines for apply-mapping are included.
One is a basic use case when there is no extra mapping file.
The other is more elaborate with several components and
a mapping file.

Signed-off-by: Martin Malina <mmalina@redhat.com>