fix(RHTAPBUGS-254): Add version 0.11 ec task update (#116)
Bump the tag used in the ec task bundle the the latest.

The specific motivation for doing this now is this test failure:

    invalid uri: uri missing scheme prefix

The failure is related to this unmerged chains upstream PR:
https://github.com/tektoncd/chains/pull/792 which has an ec workaround in
https://github.com/enterprise-contract/ec-cli/pull/649 .

IIUC updating the task bundle should mean the version of ec used includes this
workaround should get the tests passing.

There are other changes to the task definition which might cause breakages in
other ways, so we'll need to monitor the next test run and see how it goes.

Note that there are over two months of changes since the last update. To see
everything that changed we can look at the diff in the ec-cli repo, e.g. `git
diff 7f2e25e cf58b68`. In particular there are changes to the steps and the
step names.

Signed-off-by: Simon Baird <sbaird@redhat.com>