feat(HACBS-2222): add test for failing apply-mapping
Adding the capability to write test pipelines
where the task under test is expected to fail.
This is done using a new annotation added to
the pipeline object. See CONTRIBUTING.md
for documentation.

Also, adding a new test using the above
that verifies that the apply-mapping task
will fail given failOnEmptyResult=true
and a resulting snapshot with 0 components.

Signed-off-by: Martin Malina <mmalina@redhat.com>