fix: wrong files pattern for task tests workflow (#125)
The workflow is supposed to pick up versioned task dirs
(the directories that contain the task definitions)
that were changed and to run tests for each of them.

Before this fix, it could also pick up a directory one
level up. E.g. if catalog/task/kubernetes-actions/OWNERS
was changed, it would include catalog/task/kubernetes-actions
in the list of changed directories.

Signed-off-by: Martin Malina <mmalina@redhat.com>