feat(HACBS-2290): collect-extra-data stores CRs in a workspace
This commit creates the collect-data task, which replaces the
collect-extra-data task. In addition to the rename, the task now stores
the release CRs as json in a workspace so that other tasks can use them
without having to pass them as task parameters, which runs the risk of
hitting character limits.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>