feat(HACBS-2387): use cosign copy skopeo copy in push-snapshot
In order to ensure that all artifacts related to the image are also
released, cosign copy should be used instead of skopeo copy. This change
also drops some flags that don't exist for cosign copy, but did for
skopeo copy. The functionality seems to be the same though. Also, the
image for the task was updated to the release-utils image as that one
has the cosign command.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>