feat(HACBS-2405): add tests for push-sbom-to-pyxis
This change adds an optional argument to the task. With this,
you can define an init script from the utils image that gets
sourced at the beginning of each step that needs some mocks.

Mocks are needed when we want to test tasks which call external
services. In this case we want to mock calls of cosign and
upload_sbom.

Signed-off-by: Martin Malina <mmalina@redhat.com>