refactor(HACBS-2456): use git resolvers instead of bundle resolvers (#163)
This commit changes the sample for the latest version of each task to
use the tekton git resolver instead of the bundle resolvers. It also
creates a new version of each pipeline that calls its tasks with git
resolvers instead of bundle resolvers. The verify-enterprise-contract
task was left as a bundle resolver for now, waiting for the EC team to
provide a solution to moving it to a git resolver.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>