fix: improve reliability of test_tekton_task script
This script runs as part of PR checks on Github
and sometimes it gets stuck on this command:
`tkn pr logs $PIPELINERUN -f`

"-f" stands for follow. It seems sometimes it has issues
following the logs - it does and then gets stuck.

This should avoid the issue, but at the price
of not being able to see the log live as
the PR runs. It will be spit out once the PR finishes.

Signed-off-by: Martin Malina <mmalina@redhat.com>