feat: add subdirectory param to create-internal-requests
At first I thought this task was fine since it uses a unique
identifier when creating the results file. But the issue
is that this is used in the fbc-release pipeline and at
the end there is a cleanup. And currently it would
delete the whole internal-request directory - we only
want to delete data from that one pipeline run. To do that,
we need to use the subdir.

Also, use this new version in the fbc-release pipeline.

Signed-off-by: Martin Malina <mmalina@redhat.com>