fix(RHTAPBUGS-738): fbc release pipelinerun fails on task add-fbc-contribution-to-index-image (#182)
* fix(RHTAPBUGS-738): fbc task add-fbc-contribution-to-index-image failing

This PR changes the task step watch-internal-request-status to check
for .status.condition[].reason instead of .status.condition[0].status
to determine if the request has Succeeded of Failed.

The change is necessary due to the update of the
internal-services-controller to use the operator-toolkit. The older
version use to have the status as Unknown at start. After finishing
the run, the internalrequest reason was set to Failed or Succeeded
and status False or True (respectively).

The new version sets the reason to Running and status to False at the
start (causing the bug) and sets status to True and reason to Succeeded
or Failed at the end of the run.

Signed-off-by: Leandro Mendes <lmendes@redhat.com>