feat(HACBS-2405): add tests for push-sbom-to-pyxis (#162)
These tests need some mocking.

Mocks are needed when we want to test tasks which call external
services. In this case we want to mock calls of cosign and
upload_sbom.

To achieve that, a new file
with mock function definitions is added to the tests/
directory. The mocks are then injected at the beginning
of the task's scripts.

Signed-off-by: Martin Malina <mmalina@redhat.com>