feat(RHTAPREL-628): add task to collect Pyxis params
This commit adds a new task `collect-pyxis-params` to retrieve the pyxis
server and secret from the data file. It then exposes those as results
so that the pyxis tasks can use them, instead of relying on them to be
passed as pipeline params. As a result, the rhtap-service-push and
push-to-external-registry pipelines were modified to use this new task.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>