fix: bump collect-data image for bug fix
Bump the image used in the collect-data task to get a bug fix in the
merge-json script
(https://github.com/redhat-appstudio/release-service-utils/pull/67) and
a change to get-resource util to return {} instead of the empty string
if the json key doesn't exist
(https://github.com/redhat-appstudio/release-service-utils/pull/68).
Also removes all references to extraData in the task
and replaces it with data to align with recent API changes.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>