feat(RHTAPREL-618): add rhPush and commonTag to create-pyxis-image
These new parameters will be used in the
rh-push-to-registry-redhat-io pipeline.

rhPush, if set to true, will set the proper
`registry` and `repository` values when creating the Container
Image object in Pyxis. Also, the image will be marked as published.

commonTag, if set, will be used for the `tag` value in the Container
Image object in Pyxis.

This change depends on
https://github.com/redhat-appstudio/release-service-utils/pull/76

Signed-off-by: Martin Malina <mmalina@redhat.com>