fix(RHTAPBUGS-973): translate image references when signing rh images
When creating a signing request, the image reference
needs to be translated to the public facing one
which will be used for pulling of the image.

Before this change, the signatures would be requested
for the actual quay.io references. Now they will use
registry.redhat.io or registry.stage.redhat.io references.

Signed-off-by: Martin Malina <mmalina@redhat.com>