feat(RHTAPREL-619): add prepare-fbc-release task to support-multi-comp
This task introduced to support multi-component release for FBC
in this task we fetch containerImage from snapshot and collect
valid OCP tag to be used, and update fromIndex, targetIndex and
binaryImage with the found OCP tag from containerImage and
also adds the value to snapshot respective to components,
so this can be used by other tasks for
multiple component FBC release.

Also this task will results in removing get-ocp-version,
update-ocp-tag and validate-single-component once we
have the compatibility for multicomponent support.

Signed-off-by: Happy Bhati <hbhati@redhat.com>