feat: add option to run single task tests
When running task tests using the test_tekton_tasks.sh script,
it will now be possible to provide either task dirs or single test
yaml files. This will be useful for when you're working on a single
test and want to try it out - until now you would always need to wait
for all the other tests to run too (unless you temporarily renamed
or deleted the other tests).

Signed-off-by: Martin Malina <mmalina@redhat.com>