feat(RHTAPREL-762): support multiple floating tags
This commit modifies the push-snapshot task to accept a list of
floatingTags in the data.images key of the RPA instead of just a single
floatingTag string. It also modifies the push_image function to compare
the source and destination digest before every push, instead of only
checking the destination digest with the default tag.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>