feat(RHTAPSRE-287): use new group name in codeowners (#343)
A new GitHub group named `konflux-release-team` was created to replace
`Release Team`. Use new one to be able to delete former one.

The context of this new group name is to align with new service name but
also with the name of existing Rover group. Having both GitHub and Rover
group names aligned will allow to have clusters auth with either GitHub
or Red Hat SSO without having any difference in the RBACs.

Signed-off-by: Hugo Arès <hares@redhat.com>