feat(EC-272): check that operator package name is allowed
This checks that the FBC pipeline can only publish operator packages
which are declared explicitly on the ReleasePlanAdmission.

The purpose of this change is to prevent two different teams from
publishing the same operator package name to the same index, stomping on
each other. That uniqueness requirement is enforced elsewhere, where we
ensure that all RPAs in a managed workspace have no conflicting
allowedPackages values.

Signed-off-by: Ralph Bean <rbean@redhat.com>