feat(RHTAPREL-813): add rh-advisories pipeline (#373)
This commit adds a new pipeline, `rh-advisories`. It is a copy of the
rh-push-to-registry-redhat-io pipeline, but with the advisory pieces
added in. The goal is for this pipeline to eventually replace the
rh-push-to-registry-redhat-io pipeline. We are making a copy as its own
pipeline for now as to not break users (new things are required in their
CRs to complete successfully with the advisory pieces added in).

Signed-off-by: Johnny Bieren <jbieren@redhat.com>