refactor(KFLUXBUGS-1115): modify create-pyxis-image results
This commit modifies the create-pyxis-image task to no longer have
containerImageIDs as a result. These are stored in the pyxis.json file
it now creates and no pipeline is using them. It also update the
pyxis.json to include the containerImage value in each entry. Finally,
the pyxis.json file is stored in a subdirectory and the path to it is
output as a task result.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>