feat(RELEASE-266): use digest-based tag for source container origin
The build service is switching to a new tag for the source container,
so we need to use the new one which is based on the binary image digest.
(For now they push both.)

Also, push this new digest-based tag to the destination repo as well.

Also, fix a bug where the digest comparison was wrong when pushing
source containers, so the push would happen even if the destination
already had the image.

Signed-off-by: Martin Malina <mmalina@redhat.com>