feat(RELEASE-266): use digest-based tag for source container origin
The build service is switching to a new tag for the source container,
so we need to use the new one which is based on the binary image digest.
(For now they push both.)

This modifies the rh-sign-image task so that it uses
the new source container origin tag when running skopeo
to get the source container digest.

Signed-off-by: Martin Malina <mmalina@redhat.com>