refactor(RELEASE-384): push-snapshot accepts tags per component
This commit refactors the push-snapshot task to accept tags in each
individual component in the mapping. It also accepts a defaults section
in the mapping that will be added to each individual component's list of
tags.

The legacy code using data.images.floatingTags and things like
data.images.addSourceShaTag is left in the task with this commit. It
will continue to be used if no tags are provided in
data.mapping.components or defaults. Once all users are migrated, the
legacy code will be removed.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>