refactor(RELEASE-384): rh-sign-image looks at mapping (#411)
This commit changes the rh-sign-image task to take the
pushSourceContainer values from mapping.components and mapping.defaults
into account. If a value exists in mapping.components for the component
being processed, that value is used. Otherwise, it looks at
mapping.defaults. Neither is considered, however, if it is set in the
normal spot of images.pushSourceContainer. The
images.pushSourceContainer support will be removed in a future commit.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>