refactor(RELEASE-920): drop images support in rh-sign-image
This commit changes the rh-sign-image task to no longer use the
data.images key for pushSourceContainer and floatingTags. They have
been moved to data.mapping.defaults or are in each component. This also
includes dropping the commonTags parameter from the task, as they are
read from the snapshot spec file now.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>