fix(RELEASE-929): push-rpm-manifests-to-pyxis supports multiarch (#422)
This commit changes the push-rpm-manifests-to-pyxis task to properly
work with multi arch images. It does this by fetching the sbom per arch
instead of just once per component.

As part of this, the create-pyxis-image stores the os value for the
image in the pyxis.json file. This replaces the containerImage, which
was already in the file in each component.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>