feat(RELEASE-930): tekton task tests always run
This commit changes the tekton task tests github workflow to run on all
pull requests instead of if only the tasks dir was changed. If the tasks
dir was not changed, it will just exit 0 without running any tests. This
allows us to make the check mandatory for all PRs.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>