feat(RELEASE-878): collect-data creates result dir
This commit modifies the collect-data task to create a results dir in
the workspace. This is so that other tasks can write their artifact
files to this directory and we do not need to copy the creation of the
directory in case it does not exist to every other task.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>