feat(RELEASE-937): push-snapshot writes a results file
This commit modifies the push-snapshot task to write information about
each image pushed to a results file. The information per image consists
of the component name, arch, shasum, and the urls it was pushed to.

The pipeline definitions are also updated due to the push-snapshot task
requiring the new resultsDirPath parameter.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>