fix: oras commands to not hide error output
I believe the redirection of standard error output was done
by mistake, probably assuming that it might polute
the variable we're assigning to. But that's not how it works
 - stderr is not used for the assignment.

And recently a user saw a failure of the oras command, but
there was no error output, so it was impossible to figure
out what happened.

Signed-off-by: Martin Malina <mmalina@redhat.com>