fix(RELEASE-994): incorrect handling of multiarch images with 1 arch
This was caught in a test and it's a valid use case that we should
support - you could have a multiarch image (with a manifest list)
that will only contain one arch image inside.

Previously, we would handle such image as a normal single arch
image and the script would fail.

Signed-off-by: Martin Malina <mmalina@redhat.com>