feat(RELEASE-1033): add support for dockerfile in create-pyxis-image
If we're able to pull the dockerfile oci artifact, we will
add it to Pyxis when creating the ContainerImage object.

At least initially, most builds will not produce this artifact,
so if the oras pull command fails, silently ignore it.

If oras pull succeeds, but no Dockerfile is saved, the task
will fail as that would suggest that the artifact was produced
incorrectly at build time and it should be investigated.

Signed-off-by: Martin Malina <mmalina@redhat.com>