feat(RELEASE-570): add option to make release repositories public
This adds two new tasks to the `rh-push-to-external-registry`
and `push-to-external-registry` pipelines:
- `collect-registry-token-secret`: This will collect the
  `.registrySecret`
  value from the data file and pass it as a result.
- `make-repo-public`: This will use the token value stored in
  registrySecret
  to make API calls to quay.io to make repositories public.

All of this will be optional. If `.mapping.defaults.public` is `true`
or any `.mapping.components[*].public` is true, we will require
the secret name to be defined in `.registrySecret`. Otherwise
this will be skipped.

The secret needs to contain the key `token` containing the quay token.
And the token needs to have permission to call the
`quay.io/api/v1/repository/<repository>/changevisibility` endpoint for
all the repositories set to be made public.

Signed-off-by: Martin Malina <mmalina@redhat.com>