refactor(RELEASE-1024): update checkton version and allow it to fail
Originally, our goal was to write a special
script to parse the sarif file and print a nice summary.

But since then, Adam improved the output of the checkton
action, so that it has all we need (including link to wiki
for each issue), so we can just use that.

Signed-off-by: Martin Malina <mmalina@redhat.com>