feat(CLDX-78): add windows binary signing step
This step is part of multiple commits for the sign-binaries task.
Specifically, this is the Windows signing step. The rest of the task
and steps will be implemented with CLDX Jiras as marked in the code
comments.

Currently, no pipelines call this task yet, so the TODO code will not
cause any failures as this time.

Signed-off-by: Scott Wickersham <swickers@redhat.com>