feat(RELEASE-1116): timestamp variable is for build date
This commit modifies the apply-mapping task to treat the timestamp user
provided variable as the build timestamp instead of the current time. It
is found by inspecting the image's labels. The existing timestamp
functionality now exists under the release_timestamp variable.

Also, if a tag is translated with a null value, the task will now fail.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>