feat: nicer display for fbc pipelines
Functionally, this was all doing the right thing before. Tasks later in
the pipeline wait automatically for preceding tasks if the output of a
preceding task is used as a param of a later task.

However, our UI doesn't render the pipeline nicely if we only rely on
those dependencies. By using `runAfter`, we can hint to the UI how the
pipelinerun should be rendered.

Signed-off-by: Ralph Bean <rbean@redhat.com>