fix(RELEASE-1187): error when Dockerfile has a different filename
We assumed that the Dockerfile in the *.dockerfile oci artifact
would always be called Dockerfile. Wrong assumption.
We created STONEBLD-2822 for the build team to always use
the same name (instead of just copying whatever the user has).
But for now let's read the filename from the metadata and use that.

Signed-off-by: Martin Malina <mmalina@redhat.com>