feat(CLDX-82): add checksum signing step
This commit adds a step to the sign-binaries task to generate the
sha256 sums of all the signed binaries and collect them in a single
file (sha256sum.txt).
It then transfers the file to a remote host to be signed with Red
Hat's GPG key. The detached signatures are returned and released
alongside binaries in subsequent tasks to the RH Developer Portal.

Signed-off-by: Scott Wickersham <swickers@redhat.com>