fix: tweak Promote branch inputs
When you trigger the workflow via Github UI, it will
not show you the name of those inputs. It just shows
checkboxes and the description next to them.

So let's add the input names to the descriptions
and also remove the "If set to true" part, as it doesn't
make sense in the UI.

Signed-off-by: Martin Malina <mmalina@redhat.com>