fix: tweak github actions (#604)
* chore: update checkout action to v4

Some places were already on version 4. Move the rest
as well.

v3 would print this warning:

The following actions use a deprecated Node.js version and
will be forced to run on node20: actions/checkout@v3.

Signed-off-by: Martin Malina <mmalina@redhat.com>

* fix: tweak Promote branch inputs

When you trigger the workflow via Github UI, it will
not show you the name of those inputs. It just shows
checkboxes and the description next to them.

So let's add the input names to the descriptions
and also remove the "If set to true" part, as it doesn't
make sense in the UI.

Signed-off-by: Martin Malina <mmalina@redhat.com>

---------

Signed-off-by: Martin Malina <mmalina@redhat.com>