fix: fail build timestamp expansion for oci artifacts with no labels
There is no build timestamp or created field for some artifacts (like
disk-image OCI artifacts and models).

When "build_date" was equal to the empty string, this would cause the
build timestamp to be produced like this:

  ❯ date -d "" '+%s'
  1728532800

... which is always rounded to midnight on the given day.

This would cause multiple releases in a row to incorrectly assign the
same timestamp, even if the artifacts were different, had different
digests, and had really been built at different times.

The new check here will cause the behavior for date-less OCI artifacts
to align with the behavior for date-less images. It will fail the
release if the user attempts to use `{{ timestamp }}` and no build
timestamp can be found.

Signed-off-by: Ralph Bean <rbean@redhat.com>