fix: gitlint installation
Latest ubuntu does not allow system wide pip install.
It can be overriden with `--break-system-packages`,
but it's better to create a virtual env.

Note: This came up in another repo where we use ubuntu-latest.
Here were using an older version of ubuntu, but we it's probably
best we switch to ubuntu-latest as well.

Signed-off-by: Martin Malina <mmalina@redhat.com>