fix: get-advisory-severity properly uses overall impact
This commit fixes a bug in the get-advisory-severity internal task. If a
component from the releaseNotesImages matches one from the CVE and it
has impact "", then the overall CVE impact should be used. Previously,
it would only use the overall CVE impact when the component exists in
the CVE if the ps_component data was missing the impact key entirely.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>