chore(RELEASE-1334): remove github workflow pushing to internal
The last PR has merged so that the `internal` directory is removed. We
only pushed to the `internal` branch so that updates to the `internal`
dir were on the staging sre clusters. Since we will have no more updates
to the `internal` dir, we no longer need this.

Once the PR removing the final resource from the `internal` dir makes it
to the `production` branch, a second PR will be opened to update the
`promote_branch.sh` script to stop pushing to `stable`, and the two
branches (`internal` and `stable`) will be removed.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>