refactor: update existing internal-services manager prs
This commit changes the task used in the internal-services final
pipeline to update an existing PR to the stable branch if one exists
instead of always opening a new one from a new branch. This helps
prevent an endless number of branches/prs being created if the PRs
aren't timely merged.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>