fix(RELEASE-1543): add author to push-artifacts-to-cdn
The push-artifacts-to-cdn-task internal task has a required parameter
`author` but the internal pipeline that calls it did not pass it. So,
the task always failed. This commit updates the pipeline to accept the
`author` parameter and pass it to the task. It also updates the managed
task that calls the internal pipeline to retrieve the author from the
release json file and pass it to the internalRequest.

Finally, it adds some missing parameters to related READMEs as they were
listed in the yaml files but not READMEs.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>