fix(KONFLUX-8071): migrate blob-signing-pipeline
This commit migrates the blob-signing-pipeline from gitlab to this repo.
This allows us to use a git resolver to call it, instead of relying on a
cluster ref. The managed task is updated accordingly.

This is the last thing using a cluster reference, so switching this
allows us to turn off support for those in the internal-services
operator manager.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>