feat(RELEASE-1536): check for advisory id before creating it
Before creating a new advisory in the Gitlab repo, make
sure that there isn't an existing advisory with the same
year and live id. If we find one, exit with a failure result.

This check will happen regardless of whether we get the live id from
Errata Tool API, or it's set in the data file.

Signed-off-by: Martin Malina <mmalina@redhat.com>