chore: disable Mintmaker/Renovate on staging branch
When we needed to set up periodic e2e runs for the catalog
staging branch, the staging branch was onboarded to Konflux
as a new component. Because of that, we now also get Mintmaker
PRs for that branch. But those are duplicate PRs, because
we already get those for the development branch and every
Wednesday all the changes from development are pushed
to staging as part of our branch promotion process.

So let's add a renovate config to ignore the staging branch.

Signed-off-by: Martin Malina <mmalina@redhat.com>