feat: support compute resources removal in task tests
This adds a new --remove-compute-resources cli option
for test_tekton_tasks.sh. This is useful when you have
a local kind cluster with very limited resources.

Before we started adding computeResources everywhere,
I never had issues running tests locally, but now
I often run out of resources.

I was considering removing this always, but then we
wouldn't be able to catch typos in the resource
settings for example. So it's only an optional
flag that you can use for testing locally,
but in Github PR checks it will always run
with the computeResources set.

Signed-off-by: Martin Malina <mmalina@redhat.com>