fix(RELEASE-1715): argument list too long in filter task
The filter-already-released-advisory-images task is failing
on argument list too long if the snapshot is very big
(~200 components). The reason is that the full snapshot spec
json is being passed to the internal-request script as
a parameter and that can exceed the overall limit the OS
has when creating a new process.

The solution is to compress the snapshot before passing it
(using gzip and base64) and then decompress in the internal
task. In the failing case, the original snapshot spec json
string was ~160k long. After the compression, it is 29k.

Signed-off-by: Martin Malina <mmalina@redhat.com>