fix(KONFLUX-8821): increase the batch size for signing
When we are submitting larger Snapshots to be signed, we can improve
efficiency by increasing the batch sizes. This will enable more requests
to be processed at once. Since we are doing more with each request,
however, we also need to ensure that we allow for a longer timeout.

Signed-off-by: arewm <arewm@users.noreply.github.com>

rh-pre-commit.version: 2.3.2
rh-pre-commit.check-secrets: ENABLED