fix(RELEASE-1748): use floating tag as source for index image copy
This commit updates the publish-index-image managed task to pass the
index_image (floating tag) value instead of index_image_resolved (image
via digest) to the publish-index-image internal pipeline. This improves
things because the digest may no longer exist if a new push overwrites
it.

This commit also removes the sourceIndex parameter from the
publish-index-image managed task because it was not used.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>