refactor(RELEASE-1003): remove unnecessary pipelineRun.uid params
This commit cleans up tasks that have a a parameter with value
context.pipelineRun.uid passed to them in the pipelines they are used.
In each case, we can get the pipelineRun uid from another parameter, so
it is a redundant parameter (and makes the task more error prone - it
relies on the proper value being passed). There were three instances of
this:

The first was the extract-binaries-from-image task. The `subdirectory`
parameter existed with the uid value. We can get the uid from the
snapshotPath parameter instead.

The second is the push-rpm-to-koji task. Same situation as
extract-binaries-from-image.

The final one is the push-oot-kmods pipeline passed the pipelineRunUid
parameter to the sign-oot-kmods task, but the task doesn't even define
this parameter.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>