feat(RELEASE-1785): use sparse checkout in create-advisory-task
As the advisories repo grows bigger we started seeing performance
issues when cloning the repo.

New version of the git_clone_and_checkout function in the utils
image supports sparse-checkout. This will make it much faster
and use less resources.

This relies on
https://github.com/konflux-ci/release-service-utils/pull/491

A similar change was done previously for the filter task:
https://github.com/konflux-ci/release-service-catalog/pull/1287

Signed-off-by: Martin Malina <mmalina@redhat.com>