fix: gitlint run as part of merge queues
github.event.pull_request.base.ref only works for PR
events, but it's undefined in a merge queue. As a result
the gitlint check would fail there with:

'origin/..HEAD': unknown revision

The extra fetch may or may not be needed, but it shouldn't
hurt.

Note that when the CI runs directly in the PR,
github.event.pull_request.base.ref is the same as
github.base_ref.

Signed-off-by: Martin Malina <mmalina@redhat.com>