feat(KFLUXVNGD-454): add oci version tag to apply-mapping
To pull a Helm chart, we typically need it to be tagged based on the
oci artifact version (rather than by digest or arbitrary tag).

To allow that, enhance apply-mapping task to be able to fetch the
version from the artifact metadata and populate a variable that
can be later used in RPAs templating.

Signed-off-by: Yftach Herzog <yherzog@redhat.com>